Changing int to size_t
Changing the integral type of iterations counter from int to size_t, when variable could not be negative and value could overflow size of int.
In some cases this was done also to accomodate to the return type of strlen(), which is size_t.

+++ src/functions.c
@@ -90,7 +90,7 @@ int find_restaurants() {
     CURL *curl = NULL;
     CURLcode result;
     int i, len;
-    int j, head_length, length; //variables for output formatting
+    size_t j, head_length, length; //variables for output formatting
     char location[1000];
     struct json_object *jobj, *obj1, *obj2, *obj3, *obj4, *lat, *lng; //Geocoding API json objects
     struct json_object *pl1, *pl2, *idx, *name, *price, *vic, *rate, *open, *hrs; //Google Place API json Objects
